+++ Installer.php
@@ -11,6 +11,7 @@ use Composer\Package\PackageInterface;
 use Composer\Installer\LibraryInstaller;
 use Composer\Repository\InstalledRepositoryInterface;
 use Composer\Script\CommandEvent;
+use Composer\Script\Event;
 use Composer\Util\Filesystem;
 
 /**
@@ -228,12 +229,40 @@ EOF
             rmdir($yiiDir);
         }
     }
-    
+
+    /**
+     * Special method to run tasks defined in `[extra][yii\composer\Installer::postCreateProject]` key in `composer.json`
+     *
+     * @param Event $event
+     */
     public static function postCreateProject($event)
     {
+        static::runCommands($event, __METHOD__);
+    }
+
+    /**
+     * Special method to run tasks defined in `[extra][yii\composer\Installer::postInstall]` key in `composer.json`
+     *
+     * @param Event $event
+     * @since 2.0.5
+     */
+    public static function postInstall($event)
+    {
+        static::runCommands($event, __METHOD__);
+    }
+
+    /**
+     * Special method to run tasks defined in `[extra][$extraKey]` key in `composer.json`
+     *
+     * @param Event $event
+     * @param string $extraKey
+     * @since 2.0.5
+     */
+    protected static function runCommands($event, $extraKey)
+    {
         $params = $event->getComposer()->getPackage()->getExtra();
-        if (isset($params[__METHOD__]) && is_array($params[__METHOD__])) {
-            foreach ($params[__METHOD__] as $method => $args) {
+        if (isset($params[$extraKey]) && is_array($params[$extraKey])) {
+            foreach ($params[$extraKey] as $method => $args) {
                 call_user_func_array([__CLASS__, $method], (array) $args);
             }
         }