+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 composer extension Change Log
 2.0.5 under development
 -----------------------
 
+- Enh #10: Added `yii\composer\Installer::postInstall()` method (rob006)
 - Bug #11: `generateCookieValidationKey()` now saves config file only when `cookieValidationKey` was generated (rob006)